updating user profile

+++ src/UserProfileManagement/UserProfile.Application/UserProfiles/Commands/DeleteUserProfileCommand.cs
@@ -1,21 +1,27 @@
 ﻿using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Hive.Common.Core;
 using Hive.Common.Core.Exceptions;
+using Hive.Common.Core.Interfaces;
 using Hive.UserProfile.Application.Interfaces;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 
 namespace Hive.UserProfile.Application.UserProfiles.Commands
 {
     public record DeleteUserProfileCommand(int UserProfileId) : IRequest;
 
-    public class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfileCommand>
+    public class DeleteUserProfileCommandHandler : AuthorizationRequestHandler<Domain.Entities.UserProfile>,  IRequestHandler<DeleteUserProfileCommand>
     {
         private readonly IUserProfileDbContext _dbContext;
         private readonly ILogger<DeleteUserProfileCommandHandler> _logger;
 
-        public DeleteUserProfileCommandHandler(IUserProfileDbContext dbContext, ILogger<DeleteUserProfileCommandHandler> logger)
+        public DeleteUserProfileCommandHandler(IUserProfileDbContext dbContext,
+            ICurrentUserService currentUserService, IAuthorizationService authorizationService,
+            ILogger<DeleteUserProfileCommandHandler> logger) : base(currentUserService, authorizationService)
         {
             _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
@@ -29,6 +35,13 @@ namespace Hive.UserProfile.Application.UserProfiles.Commands
             {
                 _logger.LogWarning("User Profile with id: {@Id} was not found.", request.UserProfileId);
                 throw new NotFoundException(nameof(Domain.Entities.UserProfile), request.UserProfileId);
+            }
+            
+            var result = await base.AuthorizeAsync(userProfile,  new [] {"OnlyOwnerPolicy"});
+            
+            if (!result.All(s => s.Succeeded))
+            {
+                throw new ForbiddenAccessException();
             }
 
             _dbContext.UserProfiles.Remove(userProfile);
